refactor: move crdt files into separate directory
