+++ packages/loot-core/bin/profile-sql.js
@@ -4,7 +4,6 @@ import os from 'os';
 import * as sqlite from '../src/platform/server/sqlite';
 import * as db from '../src/server/db';
 import { batchMessages, setSyncingMode } from '../src/server/sync';
-import Timestamp from '../src/server/timestamp';
 import { runQuery } from '../src/server/aql/schema/run-query';
 import asyncStorage from '../src/platform/server/asyncStorage';
 import { makeChild } from '../src/shared/transactions';
