+++ packages/loot-core/src/server/crdt/index.js
@@ -1,0 +1,4 @@
+import * as merkle from "./merkle";
+
+export { merkle };
+export { getClock, setClock, makeClock, makeClientId, serializeClock, deserializeClock, Timestamp } from "./timestamp"
