+++ packages/loot-core/src/server/sync/repair.js
@@ -1,6 +1,5 @@
 import * as db from '../db';
-import Timestamp, { serializeClock, getClock } from '../timestamp';
-import * as merkle from '../merkle';
+import { serializeClock, getClock, Timestamp, merkle } from '../crdt';
 
 export function rebuildMerkleHash() {
   let rows = db.runQuery('SELECT timestamp FROM messages_crdt', [], true);
